Move the old options generator into UI folder
